requirements.txt
# FastAPI et serveur
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Base de données
sqlalchemy==2.0.25
alembic==1.13.1

# Google Gemini
google-generativeai==0.3.2

# Traitement PDF
PyPDF2==3.0.1
pdfplumber==0.10.3

# Embeddings et RAG
sentence-transformers==2.3.1
chromadb==0.4.22
langchain==0.1.4
langchain-google-genai==0.0.6

# Utilitaires
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose==3.3.0
